feat: Moved Fab component
Change the location of the Fab component,
so that it can be used anywhere.
